Hibernate and hierarchy of products

+++ src/com/musairov/shop/repository/UserRepository.java
@@ -30,7 +30,7 @@ public class UserRepository {
             user = jdbcTemplate.queryForObject(
                     "select * from customers where login = ? and password = ?",
                     new Object[]{login, passwordEncoder.encode(password)},
-                    (rs, i) -> new User(rs.getLong("id"), login, hash)
+                    (rs, i) -> new User(rs.getInt("id"), login, hash)
             );
         } catch (Exception e) {
             System.out.println("User not found");
